add abstract sentinel aspect for extension
